Trilinos: MueLu: upping time for valgrind test
Commit 9fe6ffa7b8ad enabled some longer running nightly tests.  I
suspect these are causing the valgrind testing to exceed the current
limit of 14400 (seconds?).  So I'm upping it to 21600 (seconds?).

By the way,  the timeout is not evident on the MueLu dashboard.
Instead, it shows up at https://testing.sandia.gov/cdash/index.php?project=TrilinosDriver.
Thanks to Brent P. for pointing this out.