Ifpack2:  Fix Belos scalar type for Ifpack2::Krylov.
There is a BelosScalarType traits class for determining the scalar type
to be used by Belos for a given Ifpack2 scalar type.  This was put in a
while ago to support Stokhos UQ scalar types.  At some point Krylov was
changed to not use it, and this is now fixed.