Teuchos: Improved TEST_THROW and TEST_NO_THROW macros.
Before, these macros would not print anything useful if the code in
question threw an exception of the wrong type (in the case of
TEST_THROW) or if it threw an exception at all (in the case of
TEST_NO_THROW).  This commit adds more diagnostic messages to help users
figure out the type of the exception that was thrown, and any message it
may have (if it is an instance of a subclass of std::exception).

Regarding my previous commit to Amesos2 and Ifpack2: Andrey came and
figured out that the MueLu_UnitTests* tests were failing because of a
lingering .xml file in my build directory.  The failing tests used a
collection of .xml files that MueLu installs.  MueLu no longer installs
that .xml file.  When I deleted it and rebuilt, the tests in question
passed.  Thanks Andrey!

Build/Test Cases Summary
Enabled Packages: Ifpack2, Amesos2, Zoltan2, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => FAILED: passed=125,notpassed=4 => Not ready to push! (16.76 min)
3) SERIAL_RELEASE_SS => passed: passed=109,notpassed=0 (13.45 min)
WARNING: Forced the push!