Ifpack2: Promoted Amesos2Wrapper out of experimental code.
Ifpack2::Details::Amesos2Wrapper used to require
Ifpack2_ENABLE_Experimental = ON in order to build.  It seems to be
working well, so I have promoted it out of experimental code.  It now
only requires that Amesos2 is enabled, which it also required before.

In order to justify the promotion, I fixed the already existing unit
test for Amesos2Wrapper.  I just had to change the name of the class
being tested.  It's not a very challenging test, but it passes, so I
feel more confident about promoting Amesos2Wrapper out of experimental
code.  Furthermore, Amesos2 has its own more interesting tests, so it's
OK if Ifpack2's tests for the Amesos2 wrapper are lightweight.