Tpetra: Attempt to address build error in test (for Bug 5978).
Thanks to Brent for pointing this out.  This build error has been
showing up only for one machine on the dashboard, and only for the 11.6
release tests, not for the development tests.  It seems like a
compiler-specific error, but it's easy to fix, so I'll fix it.  I'll
first try it with the development branch, and then cherry-pick to the
11.6 release branch.

Build/Test Cases Summary
Enabled Packages: Tpetra, Ifpack2, Amesos2, Zoltan2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=130,notpassed=0 (2.95 min)
3) SERIAL_RELEASE_SS => passed: passed=108,notpassed=0 (0.41 min)
Other local commits for this build/test group: f43392c