Teuchos: Fixed XML processing.
This patch comes from Christian Trott; I am running the check-in tests
for him.  Thanks to Christian for his hard work!  This commit will
serve his effort to add a framework for formal performance tests (with
captured history, using XML as a database format) to Trilinos.

Christian found that Teuchos' XML processing code (e.g., XMLObject)
had some problems, in particular with reading back in an XML file.
For example, a file that looked like this:

<Test>
 data
</Test>

which would be read in using the following code:

XMLObject database = FileInputSource(argc[1]).getObject();
std::cout << database;

would generate the following code (note extra newlines):

<Test>

 data

</Test>

Christian found that the lines of content (e.g., the "data" line that
is not part of a tag) are read in character by character, and actually
added to a XMLObject as individial "content Lines" with each of those
being a std::string.  In the above example, while the first Object has
1 content line, the read-in Object has 7 content lines.

Chris Baker further says:

The XMLObject treatment of content is weak. It caused problems when I
was writing the parser. My recollection of the problem is that there
is no concept of ordering of content and child elements, so that all
content gets squished together.  For example:

<parent>
  hello
  <child1/>
  world
  <child2/>
  !
</parent>

can only be represented as

<parent>
  <child1/>
  <child2/>
  hello world !
</parent>

or

<parent>
  hello world !
  <child1/>
  <child2/>
</parent>

by the serializer/deserializer.  This hasn't been (in the past 11+
years) a big problem because Trilinos makes very little usage of tag
content, and what few uses there are (arrays, perhaps?) don't mix tags
with content.  Just some context. There may be an existing bugzilla
item filed against this.

Christian did a lot of work and reported that the following example
works now.  "Works" means that if read in from a file and written back
out to a file, the resulting file is the same as the original file:

<Test>
<config>
   data
   more stuff
<more>
     hallo
</more>
</config>
<config>
   data
<more>
     hallo
     more stuff
</more>
</config>
</Test>

The tree builder will now keep putting characters coming from the
parser in the same contentline until a '\n' is encountered.  The '\n'
will not be added to a contentline. Also, whitespaces will not be
added as the first character in a contentLine. When an element is
finished, trailing empty contentlines will be removed (essentially the
last '\n' did indicate to start a new contentline, but nothing was
added to it, so you want to get rid of it again).

In effect given a valid XML input, leading whitespaces of a content
line will be removed, and blank lines will be removed. After reading
everything in, line will be one contentLine of a XMLObject.

---
Mark says: I manually verified that many tests passed.  I'm forcing
the push so that I can test packages which already have failing tests.

Build/Test Cases Summary
Enabled Packages: TeuchosParameterList
Enabled all Forward Packages
0) MPI_DEBUG => FAILED: passed=1269,notpassed=1 => Not ready to push! (42.77 min)
1) SERIAL_RELEASE => passed: passed=1191,notpassed=0 (2.59 min)
WARNING: Forced the push!