Ifpack2: RILUK: Simplified apply() implementation.
The apply() and multiply() code still had some residue from the
incomplete implementation of overlap that I removed in my last commit.
This residue included "temp vectors" that were always aliases, and a
private method for "generating temp vectors" that just created those
aliases.  I removed all of this and cleaned up the implementations of
apply() and multiply().