Teuchos: Added Performance Test XML Archive funcitonality.
The is commit includes a class derived from XMLObject as well as
functions to handle PerformanceTest-Archives. The goal is to
write tests which compare performance data against a gold standard
contained in an XML file or add the data to such an XML file.

The Basic Structure of an archive looks like this:
<PerfTests>
  <hostname.sandia.gov>
    <Configuration>
      <MachineConfiguration>
        <Compiler>
          Intel C++
        </Compiler>
        ....
      </MachineConfiguration>
      <Tests>
        <TPetra::CG-Solve>
          <TestEntry>
            <TestConfiguration>
              <MPI_Ranks>
                2
              </MPI_Ranks>
              ....
            </TestConfiguration>
            <TestResults>
              <Time_Add>
                0.113566 , 0.1
              </Time_Add>
              ....
            </TestResults>
          </TestEntry>
        </TPetra::CG-Solve>
      </Tests>
    </Configuration>
  </hostname.sandia.gov>
</PerfTests>

Each machine can have multiple "Configuration"s. A "Configuration" constsis of
- One "MachineConfiguration" entry which lists compile time and hardware information
- One "Tests" entry which can contain multiple tests with different names.
Each Test can contain multiple "TestEntry"s with one "TestEntry" consisting of:
- One "TestConfiguration" entry listing runtime configurations (number of threads,
  problem size etc.)
- One "TestResults" entry which lists timeing results etc. result entries starting with
  "Time_" or "Result_" are expected to contain a value and a tolerance (val, tol) or a
  value and a lower and upper bound (val, lower, upper).

When checking whether a performance test passed with PerfTest_CheckOrAdd_Test the archive
is searched for matching hostname, matching MachineConfiguration (i.e. all entries are
the same), matching TestName and matching TestConfiguration. If none is found a new
entry is added and the result is written into the file. If an existing one is found
the TestResult entries are compared. If all values match within the (in the archive file)
specified tolerances the function returns PerfTestPassed, else it returns PerfTestFailed.

If no matching test is found the return value is one of: PerfTestNewMachine,
PerfTestNewConfiguration, PerfTestNewTest, PerfTestNewTestConfiguration, or
PerfTestUpdatedTest.

A standard machine configuration can be generated through XMLTestNode PerfTest_MachineConfig().

A usage example can be found in tpetra/test/PerformanceCGSolve/cg_solve_file.hpp.