MueLu: arcadia can now parse ML files.
For pulling out detailed timing of the RAP and prolongator smoother.
This requires the AWK script removeLines.sh to remove some unwanted timers
(matrix-matrix multiplies used to rebalance).  Just make sure the script
is in a location that's on your path (and not ".").

Also added a capability to pass in a custom parsing function, which I
use for ML.  You create this function in the timelines/labels python
file.  See MLMMtimelines.py for an example.