Stokhos:  Fix static mp::vector view spec array_type for LayoutLeft.
The static mp::vector view specialization always keeps the sacado
dimension contiguous, even for LayoutLeft.  That means the equivalent
array_type needs to put the sacado dimension as the first dimension for
LayoutLeft.  I wrote a modified AnalyzeShape (called AnalyzeSacadoShape)
that does this by examining the Layout, and fixed up the tests for it.
In the future it would be nice to incorporate this in the standard
AnalyzeShape.