Kokkos: Added memory_fence and volatile_load to the atomic API
A memory fence waits until all the threads  outstanding loads and
stores are globally consistent before allowing the thread to continue.

A volatile_load takes a pointer to a memcpy-able T and memcpy it into a
temporay and returns the temporary by value.  This avoid the requirement
that users provide a volatile copy constructor overload for memcpy-able
types.

Other changes:
  UnorderedMap now uses an equal_to comparison functor.
  Use pair as the key in the UnorderedMap in the fenl example.