Stokhos:  Added new Kokkos-ified PCE scalar type for Tpetra stack.
This duplicates much of the previous work with the MP::Vector scalar
type for a PCE scalar type with dyanmic storage.  All of the Tpetra,
Belos, Ifpack2, Amesos2, and MueLu adapters were implemented along with
ETI and tests.  The ETI needs be redone to be more efficient,
particularly for CUDA.  All of the tests pass in parallel and serial for
threads and cuda.  The only real caveat to this is I have a cheesy
implementation of the division operator which doesn't work well, even
for preconditioning.  That's something we can address over the summer.