Sacado:  Fix small heap leak for DynamicStorage.
A previous commit (aee489839aad) fixed the copy constructor for all of
the Fad types to use a more optimized copy constructor path, that
ultimately resulted in calling a slightly different get_and_fill()
function in DynamicArrayTraits.  That version didn't check if the size
of the derivative array was zero or not, and was calling operator new(0)
if the derivative length was zero.  But then at a higher level I check
if the size is zero in the destructor, and thus in this case I was
never calling operator delete.  This results in a small heap leak
proportional to the number of empty Fad objects created.