Tpetra: Fixing bug in CrsMatrix::getAllValues
Turns out that the getCrsGraph() / getGraph() is the cannonical way
to adjudicate between myGraph_ and staticGraph_.  We should use it always.

Props to Jonathan for running Nalu for me repeatedly in an epic telecon
for us to track this problem down.

Build/Test Cases Summary
Enabled Packages: Tpetra
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=653,notpassed=0 (32.91 min)
1) SERIAL_RELEASE => passed: passed=581,notpassed=0 (16.11 min)