Tpetra: Map (Kokkos refactor): Implemented getDeviceView() method.
The getDeviceView() method is templated on the output Kokkos device
type OutDeviceType.  It returns a view of the underlying Map
implementation (Tpetra::Details::Map) suitable for execution on
OutDeviceType.

We may want to change the name of this method to match that used by
other Tpetra objects.  Perhaps getLocalView(), as in
Tpetra::MultiVector, would be more appropriate.

I also changed how the Kokkos refactor version of Map initializes its
internal implementation object.  Now, if it doesn't need to create the
device version of the Tpetra::Details::Map, it only creates the host
version, leaving the device version to be initialized on demand in
getDeviceView().  This ensures that Map's host-device methods work
(since they can't do on-demand initialization), but avoids copying
data to device unless needed.  I had to add an initialized flag to
Tpetra::Details::Map to make this work correctly, but that's a good
idea anyway, in order to distinguish an uninitialized
Tpetra::Details::Map (e.g., created via its default constructor) from
an initialized one.

There is a slight problem: users today prefer to pass Map around as
const, but if getDeviceView() caches like this, it can't be a const
method.  We can always change this later.  It's harder to go from
const to nonconst than it is to go in the opposite direction.