Ifpack2: LocalFilter: Impl'd getNumEntriesInGlobalRow & getGlobalRowCopy
LocalFilter implements Tpetra::RowMatrix, so it should also implement
getNumEntriesInGlobalRow and getGlobalRowCopy.  However, the current
implementations of these methods just throw an exception.  RILUK needs
them in order to accept a LocalFilter, which it must do in order to
work if the input matrix is distributed over multiple MPI processes.
This commit adds the missing implementations.

This will move towards fixing a number of Ifpack2 bugs, such as 5992
and 5987.