Panzer: Adding option to force exodus fields to lower-case
The lower-case was the default. The new default is to leave the
fields alone (and even allows spaces in the field names). The old
capability can still be turned on by specifying "User Lower Case"
in the input deck. Note that the previous change to STK was required
to achieve this.