Ifpack2: AdditiveSchwarz: Replaced loops over Y with deep_copy.
See my previous commit, where I replaced a loop over all entries of a
MultiVector, with a view.  This commit is similar, except that I also
had to add a new deep copy.  I use the new nonmember function
deep_copy for that.  Yay for making the code shorter and
thread-parallel at the same time!