MueLu: TentativePF: make new tentative factory use LIDs
In some situations it is possible to skip the GID->LID conversion, and
simply deal with LIDs. It requires to have the same GIDs in the same
order for a couple of maps, but I added an explicit check for that. If
it is not satisfied, we fall back to the conversion.