STK: Fix stk_util source issues discovered in Trilinos build
While adding files to the stk_util build for Trilnios a couple issues with the
source was noticed. There were a few places where a necessary header was not
being include, in those cases the header was included.

There were also a few places where a namespace stk was necessary to correctly
find symbols, in those cases the namespace stk was added.

(From Sierra)