Xpetra: CrsMatrix TpetraDeepCopy test: Expanded to find Vector bug.
Xpetra's (CrsMatrix,TpetraDeepCopy) test fails when using the new
Kokkos Refactor version of Tpetra.  I expanded the test to find out
the Xpetra::Vector::operator= is not doing a deep copy correctly, as
it should.  This appears to have nothing to do with Xpetra::CrsMatrix.
I'm committing the changes to the test, to help find the bug.

NOTE: These changes cause the test to fail, for either the "classic"
or Kokkos Refactor versions of Tpetra.  However, this failure exposes
a possible bug in Xpetra, so I am commiting and pushing these changes
nevertheless.