MueLu: Ifpack2: fix Additive Schwarz w/ ILU
Fixing MueLu's interface to Ifpack2::AdditiveSchwarz, so that ILU(k)
with nonzero overlap between subdomains works again.
Adding XML deck to test this.
Fixing the EasyParameterListInterpreter to accept the SCHWARZ.

The issue is how MueLu tells an Ifpack2 smoother that an initial guess
is zero.  This is important because
    1) The smoother might want to avoid unnecessary work.
    2) The smoother might solve a residual equation for a correction
       and then update the initial guess.  This makes sense in the case
       of incomplete factorizations and direct solves, for example.

This happens during the solve, so the AMG hierarchy and smoothers are
already constructed.

Previously, MueLu was calling Ifpack2::Blah::setParameterList() inside
the smoother apply, which would set an initial-guess-is-zero flag in
the Ifpack2 smoothers Chebyshev and Relaxation.

The behavior of Ifpack2::AdditiveSchwarz::setParameterList() is
different, in that calling it destroys the subdomain (aka inner) solver.
(This behavior is documented but a departure from what AdditiveSchwarz
previously did, and what other Ifpack2 solvers currently do.)  The result
was an exception at the first smoother apply.

AdditiveSchwarz can't currently take advantage of a zero initial guess,
so there's no need to call setParameterList().  However, it could, in which
case it would be nice to have some way to send this info to AdditiveSchwarz
w/o having to recompute the inner solvers.

Build/Test Cases Summary
Enabled Packages: Ifpack2, MueLu
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=24,notpassed=0 (0.89 min)
1) SERIAL_RELEASE => passed: passed=19,notpassed=0 (0.67 min)
2) MPI_SS_DEBUG => passed: passed=168,notpassed=0 (16.63 min)
3) SERIAL_SS_DEBUG => passed: passed=124,notpassed=0 (14.52 min)