Tpetra: MV (Kokkos ref.): Replaced some use of lclMV_.
lclMV_ is a KokkosClassic::MultiVector.  It is the primary
representation of the MultiVector's data in the classic version of
Tpetra.  In the Kokkos refactor version of Tpetra, lclMV_ is a
secondary view of the local data, that exists for the sake of gradual
porting.  This commit replaces some use of lclMV_ by working with
(new) Kokkos views directly.  It also adds some comments that discuss
semantic differences of methods like get1dView and get1dCopy, from
that of the classic version of MultiVector.