Tribits: Add ability for a project to change the strong warning defaults
Added two callback variables in the CallbackSetupExtraOptions.cmake file that
allows a project to set the defaults for strong warnings for C and C++. If these
variables are set it will override the Tribits provided default. If they are not
set Tribits will continue to set a default. The existing behavior was maintained
in that if only the C version is set the C++ version will be the C version plus
-Wwrite-strings. However, both of these can be overriden separately.

The variables are:

<project name>_C_STRONG_COMPILE_WARNING_FLAGS
<project name>_CXX_STRONG_COMPILE_WARNING_FLAGS

Three new unit tests were added to test the cases C options set, C++ options set
and C and C++ options are set to different values.