Tpetra: Map: Default constructor now makes nonnull Comm & Node.
(This change affects both the "classic" and Kokkos refactor versions
of Map.)  Map's default constructor is useful for giving Map view
semantics.  The default constructor used to set the Comm (MPI
communicator wrapper) and Kokkos(Classic) Node both to null.  This was
giving me problems when I was trying to implement the default
constructor for other Tpetra objects, so that they could have view
semantics too.  Rather than dig around in the innards of classes like
DistObject and CrsGraph, I thought that it would make more sense to
give the default Map a default nonnull Comm and Node.  The default
Comm is a "serial Comm" (the equivalent of MPI_COMM_SELF).

I made this change in particular to support the new
Tpetra::Experimental::BlockCrsMatrix class.  I have not yet checked in
that class, but I am testing it now.