Tpetra:  Fix setting of rowPtrs_ in the local CRS graph constructor.
The Tpetra::CrsGraph constructor that takes a local CRS graph had the
setting of rowPtrs_ commented out.  This was causing an exception to be
thrown by Ifpack2::Chebyshev in getRowInfo().  I "fixed" this by adding
a const_cast so that rowPtrs_ could be set.  This is just a temporary
workaround.