Xpetra: BlockedCrsMatrix: fixing Add function
Thanks to Ray Tuminaro for reporting this.

B.getNodeMaxNumRowEntries() may segfault due to unitilized value. But
in fact, we don't need it. The idea here was simply to have array large
enough to store a row of A. For that, we only need
A.getNodeMaxNumRowEntries(), which is ok as A is fill completed.

The second issues was that it was inserting the full array of size
maxNumEntries, instead of only the valid part of size numEntries.