Add rank-specific get_field functions that get a field by name.
These functions are members of the stk::mesh::MetaData class.

There are two variants of the API:
1. get_field(rank, name)
2. get_node_field(name), get_edge_field(name), etc.

There is also a get_side_field(name) method, where the rank of a side is known
by the MetaData class.

Added unit-tests of the get_field functions. This exposed an issue where the
internal side-rank variable wasn't being set. That is fixed now.

(From Sierra)